StateSpaceRealizables

Documentation for StateSpaceRealizables.

StateSpaceRealizables.AbstractInnerFunction

StateSpaceRealizables.AbstractStateSpace

StateSpaceRealizables.AnyRealization

StateSpaceRealizables.Balanced

StateSpaceRealizables.CompositeInner

StateSpaceRealizables.ContinuousTE

StateSpaceRealizables.DiscreteTE

StateSpaceRealizables.InternallyBalanced

StateSpaceRealizables.KautzInner

StateSpaceRealizables.LaguerreInner

StateSpaceRealizables.ProperStateSpace

StateSpaceRealizables.StateSpace

StateSpaceRealizables.StateSpace

StateSpaceRealizables.StateSpace

StateSpaceRealizables.StateSpaceRealizable

StateSpaceRealizables.TimeEvolution

StateSpaceRealizables.UnivariateInnerFunction

StateSpaceRealizables.butter_inner

StateSpaceRealizables.poles

StateSpaceRealizables.shift_basis!

StateSpaceRealizables.ssparams

StateSpaceRealizables.ssparams

StateSpaceRealizables.ssrealize

StateSpaceRealizables.ssrealize

StateSpaceRealizables.AbstractInnerFunction — Type

AbstractInnerFunction{E<:TimeEvolution}

Abstract type for representing inner functions.

StateSpaceRealizables.AbstractStateSpace — Type

AbstractStateSpace{E,T<:Number}

Abstract type for represinting systems in state-space form.

StateSpaceRealizables.AnyRealization — Type

AnyRealization <: RealizationMethod

Type for describing an uknown state-space realization method

StateSpaceRealizables.Balanced — Type

Balanced <: RealizationMethod

Type for describing balanced realizations, i.e. reachability and controllability Grammians are equal.

StateSpaceRealizables.CompositeInner — Type

CompositeInner

Type for representing compositions of Laguerre and Kautz inner functions.

StateSpaceRealizables.ContinuousTE — Type

ContinuousTE <: TimeEvolution

Type denoting that a system evolves in continuous-time.

StateSpaceRealizables.DiscreteTE — Type

DiscreteTE <: TimeEvolution

Type denoting that a system evolves in discrete-time.

StateSpaceRealizables.InternallyBalanced — Type

InternallyBalanced <: RealizationMethod

Type for denoting internally balanced realizations, i.e. reachability Grammian is identity.

StateSpaceRealizables.KautzInner — Type

LaguerreInner{E,T} <: UnivariateInnerFunction{E,T}

Type for representing Kautz inner functions.

StateSpaceRealizables.LaguerreInner — Type

LaguerreInner{E,T} <: UnivariateInnerFunction{E,T}

Type for representing Laguerre inner functions.

StateSpaceRealizables.ProperStateSpace — Type

ProperStateSpace{E,T,TA,TB,TC}

Type for representing proper state-space models, i.e:

\[dx = A x + B du, y = C x,\]

where d is a difference operator if E<:DiscreteTE, or a differential if E<:ContinuousTE.

StateSpaceRealizables.StateSpace — Type

StateSpace{E,T,TA,TB,TC,TD}

Type for representing state-space models, i.e:

\[dx = A x + B du, y = C x + D u,\]

where d is a difference operator if E<:DiscreteTE, or a differential if E<:ContinuousTE.

StateSpaceRealizables.StateSpace — Method

StateSpace{E}(A::AbstractMatrix, B::AbstractMatrix, C::AbstractMatrix, D::AbstractMatrix) where {E<:TimeEvolution, T<:Number}

Constructs a state-space model of time evolution E from the matrices A, B, C, D. The inputs are converted to a common numeric type.

StateSpaceRealizables.StateSpace — Method

StateSpace{E,T}(A::AbstractMatrix, B::AbstractMatrix, C::AbstractMatrix, D::AbstractMatrix) where {E<:TimeEvolution, T<:Number}

Constructs a state-space model of time evolution E and numeric type T from the matrices A, B, C, D.

StateSpaceRealizables.StateSpaceRealizable — Type

StateSpaceRealizable{E<:TimeEvolution}

Abstract type for representing systems that may be realized as a state-space model.

StateSpaceRealizables.TimeEvolution — Type

TimeEvolution

Abstract type denoting the time evolution of a dynamical system. See also ContinuousTime, DiscreteTime.

StateSpaceRealizables.UnivariateInnerFunction — Type

AbstractInnerFunction{E<:TimeEvolution}

Abstract type for representing inner functions.

StateSpaceRealizables.butter_inner — Method

butter_inner(η::Number, ν::Integer)

Computes a composite inner function, of time evolution ContinuousTE, with poles matching those of the Butterworth polynomial of order ν + 1.

StateSpaceRealizables.poles — Method

poles(H::AbstractInnerFunction)

Computes the poles of the inner function H.

StateSpaceRealizables.shift_basis! — Method

shift_basis(H::AbstractInnerFunction, n::Int, ω::Number)

Computes the shift basis functions up to order n evaluated at im*ω.

StateSpaceRealizables.ssparams — Method

ssparams(s::StateSpaceRealizable)

Computes the state-space parameters A, B, C, D of the state-space model s. For proper state-space models, the return value of D is nothing.

StateSpaceRealizables.ssparams — Method

ssparams(H::AbstractInnerFunction)

Computes the state space parameters A, B, C, D for the balanced realization of the inner function H.

StateSpaceRealizables.ssrealize — Method

ssrealize(H::AbstractInnerFunction)

Computes the balanced state-space realization of the inner function H.

StateSpaceRealizables.ssrealize — Method

ssrealize(s::StateSpaceRealizable)

Computes a state-space realization of the state-space model s. Returns the input if it is already of a state-space realized type, e.g. StateSpace / ProperStateSpace.